Move scylla-cql/frame/value.rs to scylla-cql/value.rs
