+++ docs/source/migration-guides/0.11-serialization.md
@@ -9,7 +9,7 @@ Before 0.11, the driver couldn't do this kind of type checking. For example, in 
 ```rust,ignore
 # extern crate scylla;
 # extern crate bytes;
-# use scylla::frame::value::{SerializedResult, SerializeValuesError};
+# use scylla::value::{SerializedResult, SerializeValuesError};
 # use bytes::BufMut;
 
 pub trait ValueList {
