+++ scylla/src/utils/pretty.rs
@@ -1,8 +1,6 @@
 use chrono::{LocalResult, TimeZone, Utc};
-use scylla_cql::frame::{
-    response::result::CqlValue,
-    value::{CqlDate, CqlTime, CqlTimestamp},
-};
+use scylla_cql::frame::response::result::CqlValue;
+use scylla_cql::value::{CqlDate, CqlTime, CqlTimestamp};
 
 use std::borrow::Borrow;
 use std::fmt::{Display, LowerHex, UpperHex};
@@ -234,7 +232,7 @@ mod tests {
     use chrono::{NaiveDate, NaiveDateTime, NaiveTime};
 
     use scylla_cql::frame::response::result::CqlValue;
-    use scylla_cql::frame::value::{CqlDate, CqlDecimal, CqlDuration, CqlTime, CqlTimestamp};
+    use scylla_cql::value::{CqlDate, CqlDecimal, CqlDuration, CqlTime, CqlTimestamp};
 
     use crate::test_utils::setup_tracing;
     use crate::utils::pretty::CqlValueDisplayer;
